% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{Organize Workflow Information}
\usage{
organize(path)
}
\arguments{
\item{path}{Path to the project directory}
}
\value{
A list of workflow information for each R script
}
\description{
This function collects all recorded workflow information within a project.
}
