% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{record_io}
\alias{record_io}
\title{Record Input and Output}
\usage{
record_io(file_path)
}
\arguments{
\item{file_path}{Path to the R script}
}
\value{
A list containing input and output information
}
\description{
This function parses R scripts and records input and output information.
}
